requirements.txt
langgraph
langchain-google-genai
playwright
python-dotenv
pydantic